+++ hi_snex/snex_core/snex_TypeHelpers.cpp
@@ -268,7 +268,7 @@ juce::String Types::Helpers::getCppValueString(const var& v, ID type)
 
 		double dValue = (double)v;
 
-		if (fmodf(v, 1.0f) == 0.0f)
+		if (fmod(v, 1.0f) == 0.0f)
 			value << String(static_cast<int>(dValue)) << ".0";
 		else
 		{
